+++ lib/flutter_native_admob_ads.dart
@@ -27,7 +27,7 @@ class FlutterNativeAdmobAds {
     required String adId,
     bool isTesting = false,
     int adsCount = 1,
-    AdRequest? adRequest,
+    NativeAdRequest? adRequest,
     void Function(String adId)? onImpression,
     void Function(String adId)? onClicked,
     void Function(String adId)? onOpened,