+++ lib/flutter_native_admob_ads_method_channel.dart
@@ -44,7 +44,7 @@ class MethodChannelFlutterNativeAdmobAds extends FlutterNativeAdmobAdsPlatform {
     required String adId,
     bool isTesting = false,
     int adsCount = 1,
-    AdRequest? adRequest,
+    NativeAdRequest? adRequest,
     void Function(String adId)? onImpression,
     void Function(String adId)? onClicked,
     void Function(String adId)? onOpened,