+++ lib/flutter_native_admob_ads_platform_interface.dart
@@ -30,7 +30,7 @@ abstract class FlutterNativeAdmobAdsPlatform extends PlatformInterface {
     required String adId,
     bool isTesting = false,
     int adsCount = 1,
-    AdRequest? adRequest,
+    NativeAdRequest? adRequest,
     void Function(String adId)? onImpression,
     void Function(String adId)? onClicked,
     void Function(String adId)? onOpened,