+++ test/flutter_native_admob_ads_test.dart
@@ -12,7 +12,7 @@ class MockFlutterNativeAdmobAdsPlatform
     required String adId,
     bool isTesting = false,
     int adsCount = 1,
-    AdRequest? adRequest,
+    NativeAdRequest? adRequest,
     void Function(String adId)? onImpression,
     void Function(String adId)? onClicked,
     void Function(String adId)? onOpened,